smart-fall-detection-monitoring-system/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
streamlit>=1.43.0
streamlit-autorefresh>=1.0.1
opencv-python>=4.10.0
numpy>=1.26.4
pandas>=2.2.2
requests>=2.32.3
mediapipe>=0.10.14
pydantic>=2.8.2
python-dotenv>=1.0.1
